python.sh: support python.exe during binary installation

+++ buildscripts/scripts/python.sh
@@ -81,8 +81,13 @@ inst=$PWD/dest/usr/local
 
 rm -f "$out"/python*
 
-cp -v python "$out/python3"
-llvm-strip -s "$out/python3"
+if [ -f python.exe ]; then
+	cp -v python.exe "$out/python3"
+	llvm-strip -s "$out/python3"
+else
+	cp -v python "$out/python3"
+	llvm-strip -s "$out/python3"
+fi
 
 # Verify that python installation directory exists and contains correct python structure
 target_lib_dir="$inst/lib/python${v_python:0:4}"
